feat: custom sort project list script menu
